Remove sketchy-null check on property existence predicates
Summary:
We generate sketchy-null errors when we're checking the truthyness of a type that can be either a falsy non-null/void value and null/void. Currently, however, we also check for sketchy-null errors when checking whether a type has a given property, using the `PropExistsP(name)` predicate, which holds on some type `t` when `t` has a property `name`.

This check signals a sketchy-null when the target type `t` may be either null or some falsy value, which isn't super useful since it only happens during a lookup on that type--it can never be the "top level" of a conditional, which is where the conflation of falsey values and undefined is most dangerous.

It looks like the *intent* behind this might have been to check whether the *property type* is null-or-maybe-falsey, but this is always checked anyways since in the one place where `PropExistsP` predicates are generated, an `ExistsP` pred on the overall expression is also generated. This redundancy can be observed in the (very small) warning diff--for each of the removed warnings, there is another warning that is not removed which reports the same thing, but with better positioning.

Reviewed By: dsainati1

Differential Revision: D18462349

fbshipit-source-id: 22143d2bfc2c63b5aeefedf7421e93ff290c23b7