[autocomplete] consolidate insertion of autocomplete magic token
Summary:
in super legacy editor integrations, the editor told us the cursor position by inserting the "magic token" `AUTO332`. this was a bit clever because it made the file more likely to parse, in the contexts where autocomplete was supported (identifiers and member expressions): `const | = ...` doesn't parse, but `const AUTO332 = ...` does.

however, this should've been an implementation detail! you can just as easily pass the (line, col) of the cursor. this is what the LSP does.

so, to avoid breaking the legacy API, this diff changes `flow autocomplete` to find and remove the `AUTO332` token, and pass its location to the autocomplete service. this lets us consolidate the implementations between LSP and legacy, and to always have a cursor location. in a followup, I plan to use the cursor location when searching the AST, instead of looking for `AUTO332`.

Reviewed By: vrama628

Differential Revision: D18507781

fbshipit-source-id: 9beb6efb769280cc6fe74df63e7758ff21ce93c9