[autocomplete] classify attempts to complete module names in imports
Summary:
if you do `import .... from 'x|'`, we should suggest module names but we don't (it's probably pretty easy, except for how many results it would return).

in the meantime, just classify them so that we know how often it happens.

this just handles `import` for now, because figuring out if a string inside a `require` call is actually a legit `require` call, not shadowed, requires looking at the file sig. doable, just later.

Reviewed By: jbrown215

Differential Revision: D18529626

fbshipit-source-id: 23b05e00128eb7a80a4e1e9c763756e804fbf381