[normalizer] Force evalutation of EvalTs under flag
Summary:
The normalizer already had an `evaluate_type_destructors` flag that,
when enabled, would attempt return the result of evaluating a type destructor.
For example, the type
```
{| ...{| x: T|} |}
```
would become
```
{| x: T |}
```
The shortcoming with the existing implementation was that it only used results
that had already been evaluated during checking. In other words it did not initiate
a type destructuring on its own. See diff in expected tests for when it wouldn't.

This diff changes this by having the normalizer call `mk_type_destructor` (of
flow_js.ml) using the existing resolved context.

There were a couple of issues to address here:

1/ Type parameters (`BoundT`s) appearing inside types. In regular checking, we
expect all these types to have substituted away and replaced with their bounds
through `generate_tests`. One of the options for the normalizer was to replicate
this behavior, but this would produce unhelpful results. Eg. for a type like
```
type S<X: {}> = {...X}
```
we would use `{}` as the bound of `x` and get
```
type S<X: {}> = {}
```
So, instead of replacing with the bounds, we chose to stall evaluation on abstract
types, and fall back to the current method of replicating the nominal form of the
Eval-type. In this case `S` would remain as is.

We achieve this by allowing `BoundT`s in Flow_js under a new checking mode
(`Context.Normalizing`). Our goal is to allow these `BoundT`s to be recorded
on the graph (through `BoundT ~> OpenT` flows) so that they can later be
reconstructed by the normalizer. All other flows of the form `BoundT ~> OpT` where
`OpT` is some destructuring operation on the BoundT are disallowed, raising an
exception that is caught in the normalizer, and the fallback method is used.

2/ Recursion on EvalT types. While not common, this is possible (see
`RecursiveTypeDestructor` for example). To avoid non-termination we use a
cache like we do for tvars.

Reviewed By: samwgoldman, mvitousek

Differential Revision: D18522734

fbshipit-source-id: fb2896c37557c11cb5f59a14f4b353a46395e501