[autocomplete] explicitly return empty results for singleton types
Summary:
there's nothing we can suggest when you do `type T = 'foo' | 'barAUTO332'` or `function f(x: 'barAUTO332') {...}`.

if you do `(y: 'barAUTO332')`, we theoretically could use the type of `y` but I think that's a very limited use to justify a bunch of complexity, so punting on it for now.

if you do `U<'barAUTO332'>`, we theoretically could suggest the upper bounds but we don't do any suggestions based on uppers today, so punting on that too.

Reviewed By: nmote

Differential Revision: D18616795

fbshipit-source-id: cbe902913a7ffe6c0113963559b618b552247d94