[enums] Support typeof and != null refinement for enums
Summary:
You can't do much with a union of an enum type and something else. This allows you to refine to the enum value using `typeof` and `{!=}= null`. This also accurately reflects what is happening at runtime.

While we do this with things like `typeof x === "string"`, we aren't actually exposing the representation type directly: after refinement it's still of the enum type, not a string.

In a later diff we will accurately model the truthiness (or lack thereof) of the representation type so we can model `if (x) { ... }` and `if (!x) { ... }`

Reviewed By: dsainati1

Differential Revision: D18609046

fbshipit-source-id: a9c466396636247175fd399b1a04c9dcf23a6a75