[optional-chaining] Evaluate subexpressions of optional chains in refined environments
Summary:
In the current design for optional chains, as we walk up the chain we generate predicates that will hold if the chain does not short-circuit (as well as ones that will hold if it does). We use these in conditionals, but they can actually be useful when typechecking the chain itself as well. Consider:
```
declare var obj: ?{val: number, fun: (number) => string};
```
What can we do if we want to call `obj.fun`, if `obj` exists, on `obj.val`? Currently we'd write:
```
if (obj) {
  obj.fun(obj.val);
}
```
That's three whole lines! Unacceptable.
We might be tempted to write
```
obj?.fun(obj?.val);
```
but `obj?.val` has type `number | void`, and `obj.fun`, if it exists, only takes `number`. However, we can realize that if `obj` is null or void, then we actually wouldn't get to the point of calling `obj.fun` on anything! We would short-circuit and no function would be called in the first place. Therefore, we can actually just write
```
obj?.fun(obj.val) // no optional chaining on the argument!
```
If we're evaluating the argument, then we haven't short-circuited, and so we can refine `obj` to being not null or void before we evaluate the argument subexpression.

This diff adds this feature, which turns out to be pretty simple--we just refine the environment of subexpressions (arguments to calls, or indexes to indexed member accesses) with the non-short-circuiting branch of the the conditional before evaluating them. We already have these refinements, if any, because we've been building them up as we go, and there shouldn't be much overhead because we only have refinements to apply when there's an optional chain operator lower on the chain.

Reviewed By: panagosg7

Differential Revision: D18483827

fbshipit-source-id: 9d3f07f1d3f350b479f63948ca27a3b745218d0d