[optional-chaining] Optional chains used for typeof, instanceof, and Array.isArray refinements
Summary:
Similar to D18458524, which added support for refinements in literal comparisons with optional chains, this diff adds support for type comparisons with optional chains. For the most part this is a simple extension of the existing work in D18458524, in order to generate refinements for cases like this:
```
class C { ... }
declare var x: ?{
  arr: Array<number>,
  obj: {...},
  c: C,
  maybe_void: number | void
};
if (Array.isArray(x?.arr)) { ... }
if (typeof x?.obj === "object") { ... }
if (x?.c instanceof C) { ... }
```

In all of these conditionals, we can be certain that if the condition returns true, then `x` is not null or undefined, because short-circuiting always produces `undefined`, and
1. `Array.isArray(undefined) === false`
2. `typeof undefined === "undefined"`
3. For all X, `undefined instanceof X === false`.

Case 2 above means that we need to flip the predicates generated from the optional chain when looking at a program like:
```
if (typeof obj?.maybe_void === "undefined") { ... }
```
In that case, we don't know whether the conditional succeeding means that `obj` is null or undefined or if `obj.maybe_void` is void--it could be either. But in the else case of the if, we know that `obj` is not null or undefined and that `obj.maybe_void` is a number.

Reviewed By: panagosg7

Differential Revision: D18496275

fbshipit-source-id: 8eb2a23fe57838ee670e2894234e75ee0dbc8cf2