Better error messages for refining read-only properties
Summary: We generate some really bad error messages when we try to refine an object as having a property with `PropExistsP` when the property is read-only. See https://flow.org/try/#0CYUwxgNghgTiAEA3W8AqARASgLngbwFowBXGOAOwBdc8B7AM3oGcRKB1AS2EoAtdziAWwBGIGAF9xAbgBQHevAAUGTADoSZEFQCU+eOJlA for an example (the first error). With this diff, we thread a reason through with the `PropExistsP` refinement so we have something better to point at.

Reviewed By: dsainati1

Differential Revision: D18495253

fbshipit-source-id: edb8de49287e8ff8b979605924329800f79d2494