[optional-chaining] Optional chains used for null, void, and literal refinements
Summary:
Similar to D18458524, this diff adds the ability to refine optional chain expressions by comparing to literals. Taking a program like this:
```
declare var obj: ?{ a: ?{b: number, c?: number} };
if (obj?.a?.b === 42) {
  ...
}
```
We know that when an optional chain short-circuits it produces `undefined`, so if the equality check passes, then the chains must not have short circuited. This means that in addition to the refinement `obj.a.b is equal to 42` that we would normally have added, we can also add the refinements `obj is not null or void` and `obj.a is not null or void` to our heap refinements. This diff adds the ability to create these sorts of refinements by hooking the `condition_of_maybe_sentinel` function, used in generating refinements on comparison operations, into the `optional_chain` function, which produces predicates based on whether optional chains short-circuit or not.

One tricky thing, which I try to discuss in the comments, comes up when we compare optional chains to null or void. If we have an expression like this:
```if (obj?.a?.c === undefined) { ... }
```
then, unlike the `=== 42` case above, we'll hit the truthy case of the `if` when the chain does short circuit, since short-circuiting produces undefined. However, in the else case of the if, we can refine `obj` and `obj.a.` to not be null or void--we flip the truthy and falsy branches of the predicates that we built when we traversed the optional chain. The same is true for an expression like
```if (obj?.a?.c == null) { ... } // non-strict equality
```
since `undefined == null`, but it is *not* true for
```if (obj?.a?.c === null) { ... } // strict equality
```
since optional chain short-circuiting always produces `undefined` (even if the short-circuiting was triggered by `null`) which is not strictly equal to undefined.

Reviewed By: panagosg7

Differential Revision: D18458524

fbshipit-source-id: 0d88e93544859d4adffe3325d8a2e56d175102fa