[optional-chaining] Optional chains generate refinements
Summary:
This diff allows refinements to be generated from testing the truthyness of optional chains.

In a program like this one,

```
declare var a: ?{b?: {c?: {data: number}};
if (a?.b?.c) {
  ...
}
```
we can consider what conditions have to be true for the expression `a?.b?.c` to be truthy. Short-circuiting of optional chains always results in `undefined`, which is falsy, so this expression must not have short circuited. This, in turn, means that `a` is not nullish, `a.b` is not nullish, and `a.b.c` is truthy. Therefore, this program is equivalent to an expanded program like
```
declare var a: ?{b?: {c?: {data: number}};
if (a != null && a.b != null && a.b.c) {
  ...
}
```
This diff implements refinement generation for optional chains like the one above, allowing users to use optional chains in conditionals and get useful refinements. The main addition is to the `optional_chain` function, which traverses chains of property accesses, and which now also (if the `predicates` parameter is true) builds refinement predicates as it goes through the chain.

When the `optional_chain` function encounters an optional chain operator `?.` in the context of a property read or indexed read, the expression to the left of the operator (the object being read from) is refined to be not null or void in the truthy branch, and if that expression is itself is a member access, its object is refined to have the property being read. In other words,
```
if (a.b?.c) {...}
```
generates the refinements `a.b not null or void` and `a has property b`, in addition to the refinements `a.b.c is truthy` and `a.b has property c` (which is what would normally/previously be generated by `if (a.b.c) {...}`).

This diff also makes some minor modifications to the predicate/guard system in general, mainly by introducing `PropNonMaybeP`, which is a predicate similar to the existing `PropExistsP`, but where `PropExistsP(prop)` checks that an object has a truthy property `prop`, `PropNonMaybeP(prop)` checks that the object has a non-nullish property `prop`.

Reviewed By: panagosg7

Differential Revision: D18355800

fbshipit-source-id: 3ab94e6e7013f70b797e9b9c37970f77d4ebbf4d