When updating `Dependency_info`, pass a map rather than another `Dependency_info`
Summary: In D18580785, I change the representation of `Dependency_info` to be more expensive to construct. Rather than construct a partial `Dependency_info` structure from which we apply the necessary updates, we can apply updates directly from unidirectional maps. This makes it so that incremental updates are not unnecessarily expensive.

Reviewed By: gabelevi

Differential Revision: D18580793

fbshipit-source-id: 06659ef62c8aa4b4c0e428ca20679d7f7311b433