Calculate partial dependency graph separately from converting to dependency info
Summary:
Later in this stack, I change the `Dependency_info.t` representation into something that is a bit more expensive to construct than a simple map from a file to its dependencies. Because it is more expensive to construct, I want to avoid constructing it when possible.

Currently, we construct a partial `Dependency_info.t` and use it to incrementally update the full `Dependency_info.t` structure. Instead, we can incrementally update the `Dependency_info.t` directly from a map that contains updates (D18580793).

To that end, I want to make it possible to get the partial dependency graph as a map, instead of forcing consumers of the `calc_partial_dependency_info/graph` function to take a fully-constructed `Dependency_info.t`. In order to accomplish that, I am making the `dependency_info_of_dependency_graph` function available to clients, and forcing them to construct the `Dependency_info.t` if they need it. This makes it simple to switch over to using a map for incremental updates here: D18580793

Reviewed By: gabelevi

Differential Revision: D18580791

fbshipit-source-id: 4434d2ff4ec8ea2c8257e87b4b23637973d8b47b