Move filtering of dependencies earlier in calc_partial_dependency_info
Summary:
It makes a little bit more sense to do the filtering as part of calculating the dependencies for a file, rather than as part of creating the `Dependency_info` structure. The only downside is that there is a tiny amount of unnecessary work done for classic mode.

But mainly, I'm doing this so that I can pull the code to create `Dependency_info` out into a separate function in the next diff.

Reviewed By: gabelevi

Differential Revision: D18580788

fbshipit-source-id: 9dc16dd804413af757645491c8c3209c00f2b672