Don't clear coverage maps on merge in types-first
Summary:
Both merge and check call `remove_old_results` on rechecks, to remove
errors stored on the server, so that new errors can be recorded. Along with
these errors, we would also clean up coverage entries for files that changed.
The problem with this approach in types-first is that we might be checking fewer
files than we are merging (due to optimizations). Since coverage is now computed
on check, clearing the coverage maps earlier (on merge) would inadvertently cause
files to be missing in the coverage report.

This diff fixes this by avoiding to remove old coverage results on merge in Types-First.
Instead it does this on check only, which is what produced coverage results anyway.

Reviewed By: samwgoldman

Differential Revision: D18642537

fbshipit-source-id: 193b23b66500278b3d0301b94f47f1db0805a226