[easy] Improve batch-coverage testing
Summary:
This improves testing on batch-coverage by:
* *also* checking in types-first and comparing the results
* adding a remove file scenario

The classic vs types-first output diff shows a bug in the logic of reusing
coverage results in Types-First. Some files are reported as having 0 coverage
after firing a recheck optimization. This is because we are reusing coverage
results we got during *merge* instead of *check*. This will be fixed next.

Reviewed By: dsainati1

Differential Revision: D18642538

fbshipit-source-id: 0cb3dab007d691b9edca5c9a3c4636354f78d220