Compute coverage only on check in types-first
Summary:
Before, we were computing coverage in Types-First twice: one time after
the merge step and one after check, with the latter results updating the
former. This recipe makes sense in Classic (where the check phase is
a no-op), but not in Types-First, for two reasons
1. if the checking step is skipped (e.g. due to some optimization), we
would reuse the result from merge, which is wrong.
2. this is unnecessary work.

With this change we only compute coverage on check in Types-First.

This change could have been combined with the next diff. I made it separate
to show the difference in the coverage reporting. Note the output where we
report coverage on fewer files after a checking opt. This is because we had
cleared out coverage entries on merge, but only checked a subset of those.
This is fixed in the next diff where we avoid invalidating coverage entries on
merge in types-first.

Reviewed By: samwgoldman

Differential Revision: D18547279

fbshipit-source-id: 64051ed93f37fe29ec0a6267c5d18a953dcb6377