Create new use_t, TypeCastT for typing casting
Summary: Explicit type casting `(x: T)` encodes some additional information: the intent of the user. We can use this information to customize the behavior of type casting in certain situations (e.g. allow for a user to cast an enum to its representation type, when usually this is not allowed).

Reviewed By: dsainati1

Differential Revision: D18574963

fbshipit-source-id: 9ef4eba43cc08101942d9832b25307c2c45c4c85