Fix order of evaluation on jsx children for correct refinement
Summary:
Flow currently evaluates the children of a jsx expression before it evaluates the expression's attributes. This is contrary to the actual behavior of compiled JSX expressions at runtime, and leads to refinements getting invalidated in the wrong order.

Compare the behavior of flow https://fburl.com/grv8wua8 with the output produced by  babel https://fburl.com/95kvrez6

This diff moves the call to `collapse_children`, which typechecks jsx children to after the checking of the attributes.

Reviewed By: gabelevi

Differential Revision: D18642975

fbshipit-source-id: 965e9a81f6e80084af9a20cdb27842a890c4edb8