Graph
Summary:
This is an immutable graph structure which maintains backwards edges. See D18580785 for the motivation.

This is a generic structure, but it's been tailored to support my use-case. In particular, there aren't the typical functions to add/remove a single edge; instead there is the batch `update_from_map` function. If needed, it wouldn't be a problem to add those functions later.

Reviewed By: gabelevi

Differential Revision: D18580792

fbshipit-source-id: b8be7f81c500c156b2de83b3fdc23d7f6bd9033e