make documentUri != string
Summary:
We have a bug where, if you have a file open at the time hh_server is started, then you see an exception report about file-path mismatch.

The problem is that we're sending a uri to IDE_OPEN, where it really expects a filename.

The root cause is that we're using `string` for both URIs and filepaths (in particular, `type documentUri = string` inside lsp.ml). That's why we didn't catch the problem.

This diff makes documentUri be its own type, not the same as string. The diff doesn't do any more than that - doesn't fix the bug yet. I'll leave that for a future diff.

I explored also using `Path.t` instead of string to indicate places where we have a filepath. But that proved too onerous. (1) ServerRPC throughout uses string for filenames, and converting back and forth between Path.t was a pain. (2) Path.t isn't just a form of string. It also comes with its own contract of always being a realpath. (3) Because of that realpath semantics it's not a lightweight dependency; it brings all of sys_utils.

Maybe we might decide to adopt Path.t in future. Maybe we might not. But whatever decision is made there, it remains useful to distinguish URIs.

Reviewed By: arxanas

Differential Revision: D18654288

fbshipit-source-id: 06c6c9282ac1e69854dabcd0138ae899c881037b