Add typing support for `symbol` type, `SymbolT`
Summary:
We add typing support for the `symbol` type. It works similarly the other primitive types like `string`, `number`, etc.

We remove the now uneeded `Mixed_symbol` variant of `Mixed`. Refinement is improved to fully support symbols. Additionally, this allows us to fully implement casting of a symbol enum to its representation type.

Note that just like `number` is not a `Number`, and vice-versa, `symbol` is not a `Symbol`. In the core lib def, we change usage of `Symbol` to `symbol & Symbol`. Eventually, it will just be `symbol`. The `symbol & Symbol` is temporary for one version to allow for existing usage of the `: Symbol` to be codemodded to `: symbol`.

Reviewed By: gabelevi

Differential Revision: D18661792

fbshipit-source-id: c8df74ac0017104827f73ced39ad68bd827ddd11