[hack] remove Utils.log
Summary: This diff removes Utils.log since it's no longer used. And it was a bit goofy - it's a global mutable function variable, which is normally a no-op, but if `Utils.profile_log` has been set, then worker init makes it print to stderr. We don't like global mutable variables if we can avoid them. And we can!

Reviewed By: CatherineGasnier

Differential Revision: D18747174

fbshipit-source-id: 323c04ab510419aa3f39e3b3ecb5ce36ea884518