default to `all: false` in tests
Summary:
as long as flow is opt-in by default instead of opt-out, the tests' default should match.

so, instead of passing `--all` on the CLI by default and turning that back off with `all: false` in `.testconfig`, now you have to set `all=true` in the test's `.flowconfig` to turn it on.

I added it to all of the existing tests that needed it, but IMO we should go through and add `flow` to those test files and remove it. I just didn't want to bump all the line numbers around.

Reviewed By: gabelevi

Differential Revision: D18783809

fbshipit-source-id: ffd7adcbf6932bf31d80e2727f06079c52b5b248