Invalidate Eval-cache on Attempted_operation_on_bound exception
Summary:
Flow_js.mk_type_destructor was not idempotent w.r.t. exceptional behavior,
because we didn't account for its caching effect.

For example, the first time we attempt to destructure `EvalT (BoundT, id)`
in normalizer mode, we would:
1. set `id` as evaluated in the eval-cache.
2. Flow `BoundT ~> OpT`, where `OpT` corresponds to the evaluator in question.
3. Throw an `Attempted_operation_on_bound` exception.

Note the tvar `tout` that holds the result of the operation will be left with no
lower-bounds.

The second time we encounter the same `Eval (_, id)`, `id` will first be looked up in
the cache. The lookup will succeed and `tout` will be returned, bypassing the
`Attempted_operation_on_bound` exception mechanism that the normalizer
depends on.

We need to avoid reusing spurious result on subsequent evaluations of the
same EvalT. This diff does this, by restoring the evaluation cache, whenever this
exception is hit during normalization.

Reviewed By: mvitousek

Differential Revision: D18797538

fbshipit-source-id: 2a65c76073b02f392199166c7c464fed494a81d4