[enums] Add new use op for compatibility checks when exhaustively checking enums in switches
Summary:
It is an error to switch on an enum, and use anything other than that enum's members in the case tests.
We previously did not supply a use op for this and that resulted in a bad error. Create a new use op here and use it, improving the error message.

Reviewed By: jbrown215

Differential Revision: D18778521

fbshipit-source-id: 1a3a4c18b06fc1079f48b0250045ca153fee6460