[optional-chaining] fix to parenthesized optional chain assignment
Summary: [this](https://flow.org/try/#0PQKgBAAgZgNg9gdzCYAoVATApgYxgQwCcswA3IsADzAC4wA7AVxhjAB8wBvMAT1q7AAvfkwC2AIyyEwAX1noAFJQD8AOh4BKVcIC8YACwAmANxA) should error. it doesn't because, in order to allow field assignments and field deletions to use the same code, i was using the `optional_chain` function to evaluate the object being deleted from/written to, and uses the filtered, non-nullable result type. this diff changes that to only be the case when it's a deletion of an optional chain; otherwise we just use the normal `expression`

Reviewed By: panagosg7

Differential Revision: D18803649

fbshipit-source-id: 365a846b2d9ce021e50a14ad70c44de8371b9ad1