[hack] End fixed point iteration early if we reach a fixed point
Summary:
As the title suggests.

When checking loops, we currently iterate `alias_depth` times, where `alias_depth` is the biggest chain of dependencies between locals in the loop. It's known that (a) this does not always result in a fixed point (e.g. you can construct ever larger generic types), and (b) very often the fixed point will be reached sooner, but we continue iterating anyway.

This diff tackles (b). We can exit the iteration early if the environment (in fact, the per-continuation environment associated with the Next continuation) remains stable.

Reviewed By: Wilfred

Differential Revision: D18833199

fbshipit-source-id: 4fad670abeb1b9ddb76a1bcfcf07ce1ac9179554