[enums] Add enums to value used as type error message
Summary:
Enums are also types.

Initially I thought about only adding "or enum" if the enums option was on, but looking at the code I don't think it's worth it to thread the options through the code. If the enums option is off, the message is not incorrect, just slightly not as relevant.

Reviewed By: jbrown215

Differential Revision: D18844442

fbshipit-source-id: 1503aa10b2fe6bb97411d7c160eb451110a6a552