add FilterOptionalT
Summary:
In a future diff we would like to decompose `UnionT X ~> OptionalT Y` constraints into the following: `UnionT X ~> UnionT (Y, VoidT)`.

An invariant of the `UnionT` upper bound speculation process, however, that the upper bound union arises from an annotation, and thus contains only 0->1 tvars. `OptionalT` types can arise from annotations, but they can also be created via a "filtering" method, wherein we flow a type to an `OptionalT` containing a tvar, to "strip off" `void` from a type, and use that tvar in place of the original type. This process is not guaranteed to produce a 0->1 tvar, and thus is not compatible with speculation.

In order to ensure that we can perform the desired decomposition, we introduce a new `FilterOptionalT` type that behaves like an `OptionalT` upper bound. The upside of this is that we now have the guarantee that `OptionalT`s are no longer created by this filtering process, and thus that they contain tvars compatible with speculation.

Reviewed By: samwgoldman

Differential Revision: D18785765

fbshipit-source-id: dda3b5a8dbf38e2a6302ce313b0bf6b026692374