[types-first] Error on exported class and function assignment
Summary:
Before, Flow classic and types-first both allowed the pattern
```
class A {}
A = 1;
export {A};
```
but the signature they produced differed: classic would include the updated
value of `A` (number), but types-first the class definition. To address this
discrepancy, we are making this pattern an error in types-first.

The error is raised in `Env.update_var` whenever an exported class or function
is being reassigned. The check needs to happen here, since the signature
extractor does not visit all statement, even though the environment of exported
bindings is populated much earlier during signature generation.

The fix for this kind of error is reassign to another local variable and export that.

Reviewed By: samwgoldman

Differential Revision: D17610639

fbshipit-source-id: 83ec6f55d94482ad5c8775c532d336c9eaff09d4