[enums] If enums option is off, don't introduce new binding, use `any` type
Summary:
Previously, if `experimental.enums` was not on, we errored at the `EnumDeclaration` site but otherwise continued as usual.
Instead, we should not introduce a new binding, and type the enum declaration as `any`.

An alternate approach would be to still introduce the binding, but type it as `any`. In any case, it shouldn't be typed as the actual enum.

Reviewed By: dsainati1

Differential Revision: D18872836

fbshipit-source-id: c29bbaa700d49eab294c0e4a0fa9b1f923131e86