Cache type_contents results to speed up LSP DefinitionRequest
Summary:
This creates a cache as part of the persistent connection client info, and uses it to speed up DefinitionRequests coming from an LSP client.

`type_contents` is a function of (a) the contents of the file being typechecked, and (b) the stored `SigContext`s in shared memory. Therefore, the cache invalidation strategy is to invalidate the entire cache on a recheck (because it perturbs the `SigContext`s, and is the only thing to do so), and to invalidate the cache for each file whenever that file changes (using the DidChange messages).

Storing the cache with the rest of the persistent connection client information makes invalidation fairly straightforward, since any DidChange message can invalidate the cache for that file. We know that the contents used to service persistent requests are only changed when DidChange messages come through.

Reviewed By: gabelevi

Differential Revision: D18739043

fbshipit-source-id: 62094eb8e36c779c25e77fc832e2c7856c566cf7