[parser] improve error handling of invalid .tree.json tests
Summary:
even with `~fail:false`, the JSON parser still raises exceptions when it encounters errors from which it can't recover (it could probably return `null` instead...).

so when a parser test has an invalid .tree.json, it can throw and kill the entire test runner. instead, we should catch the error and fail just that test.

also noticed that when it fails on invalid encoding, which happens while reading the contents of the string before parsing, we end up throwing a `Unexpected end of file` as the first error instead of `Malformed encoding`, which is hard to debug. so, now we avoid adding an extra error if there already is one.

Reviewed By: nmote

Differential Revision: D18907232

fbshipit-source-id: 18a61da2abb09447ce02951406736e7e07ee2f56