Add signatures in sharedMem.ml to NoCache, WithCache, and LocalCache
Summary:
These functors previously had no signatures for the modules they output. I think this was omitted because it was difficult to figure out what the signatures should be. I'm filling them in now. This will make it easier later to add `sharedMem.mli`.

Originally I wasn't going to deal with `LocalCache`, since it's not really exported from `SharedMem`. However, some of the tests need to deal with it, so I added a signature to it which will allow {D18926237} to export the parts needed by the test

Reviewed By: samwgoldman

Differential Revision: D18926235

fbshipit-source-id: 0747d37e3f6ec2c1be86b93229667dc09b322d25