Write properties to unsealed objects in successful speculation scenarios
Summary:
While Flow is speculating, we don't want to perform any stateful actions that might pollute the state of the checker if a speculation case fails. Most of the time, stateful action means adding a constraint to the graph, and we model this by deferring these actions until we can pick a successful speculation branch.

However, writing to an unsealed object is also a stateful action. Previously we didn't have a model for this action, so we would just not perform it during speculation. This led to weird behavior like this:
```
({} : {p : string});
```
being okay, but this:
```
({} : {p : string} | number);
```
being an error.

We can instead create a new speculation `Action` to encode writing a property to an object, and defer these writes until we successfully choose a branch.

Reviewed By: panagosg7

Differential Revision: D18820154

fbshipit-source-id: 7ff1c1794b1b4ab762d5a530e11275dabc67b90e