Delete the SQLITE3 sections of hh_shared.c
Summary:
In `hh_shared.c` there is some code which is disabled when `NO_SQLITE3` is defined. In Flow, this is always defined, so that code is dead.

This diff deletes that dead code, but doesn't delete the stubs just yet. I used the `UNUSED` macro to make the compiler happy.

Reviewed By: samwgoldman

Differential Revision: D18939394

fbshipit-source-id: 2de3c9d53812684938b6c3c648f6eb531dd9b786