allow full optimization guard on union equality checks
Summary:
We originally had a special optimization check for flows of the form `UnionT ~> EqT (UnionT)` that caught comparisons between equal enums. However, the would fail if the unions were not enums, and in the case where a union is a union of large enums, this could cause us to do a lot of extra work.

We here expand the optimization check to make use of the full optimization guard. Note that because equality checking is slightly different from subtyping, we parameterize `union_optimization_guard` over the function used to compare types.

Reviewed By: gkz

Differential Revision: D18966549

fbshipit-source-id: e4d806ffe60de90057fed0fd3c2eaee48c2ef64c