[spreads] Allow indexers that unify
Summary:
Allow matching indexers to be spread.

Note that this rule may have a confusing side effect-- if the indexers don't unify successfully, then we still take the type of the second indexer in the spread.

The most ideal way to do this would be to fully resolve the types of the indexers (possible because we don't infer any indexers today, so they must come from annotations) and then use the speculation machinery to check if they would unify. If they would, propagate the indexer, otherwise error on the spread and propagate any.

However, we don't know that we won't ever want to infer indexers, and that change is _way_ more complicated than this one. This change also does not stop us from making that improvement in the future if we would like to.

Reviewed By: dsainati1

Differential Revision: D18407544

fbshipit-source-id: 152f357c6e07cff62afdb8b05f6f57b2208ac223