[enums] Disallow strict equality checks of enums in if statements, and anywhere with non-enum values
Summary:
Background: switch tests (between switch discriminant and switch case) are modelled as `===` in Flow.

What we want to achieve:
- Ban `===` checks of the same enum in if statements
  - We want users to check them exhaustively in switch statements instead
  - In the future we may add support for exhaustively checking enums in if statements, but we want to disallow it in the mean time so that if we do implement it we won't suddenly have a lot of erroneous code
- We do continue to allow `===` checks of enums in other conditional scenarios, that will never support exhaustive checking, like loop conditions and ternary expressions, or outside conditional checks, like when used as an expression (e.g. could be used in filtering scenario `.filter(x => x === E.A)`
- Ban `===` between an enum and non-enum value (other than with `null`/`undefined` outside of switch statements so we continue being able to use this to refine away these in an optional enum)
  - This helps ensure that we aren't breaking the "opaqueness" of an enum by allowing comparisons with its representation type

I previously added some custom logic around this when doing exhaustive checks (e.g. switching on an union of enums). This diff subsumes these special cases: cleaning up code, checking cases the previous code missed, and making things consistent throughout.

Implementation:
In `statement.ml`, we previously had an `is_cond` boolean flag in a variety of places. We replace it with an option of `cond_context`. If the flag was previously `false`, we use `None`, if it was previously `true`, we use `Some` with further information on what the condition was: if, switch, or other. We track this information in `StrictEqT`.

Future work:
In subsequent diffs, I will
- Rename the `ExhaustiveCheck` use op to something more general like `SwitchTest` and refactor its language a bit, as this is more accurate (and could be used in the future for other purposes).
- Add a custom error for when we `===` and enum in an if statement
- Delete the no longer needed `EEnumExhaustiveCheckOfUnion` error

Reviewed By: dsainati1

Differential Revision: D19057875

fbshipit-source-id: 09127b30bfe9c2da043fcb96f4c408324d6b2d42