Stop mapping shared memory to a fixed address
Summary:
Now that shared memory uses relative pointers, it's not necessary to fix the
mapping to a specific address.

This might have some small effect on the TLB cache, but I'm not expecting a
noticeable perf impact from this change. Instead, the motivation here is that
mapping to a fixed address is kind of yucky.

Reviewed By: gabelevi

Differential Revision: D19183603

fbshipit-source-id: 9cfb72ceb187fcea272875d112ee71604cb1c2e2